+++ lib/ts/recipe/emailpassword/types.ts
@@ -25,7 +25,7 @@ import {
     Styles
 } from "../../types";
 import EmailPassword from "./emailPassword";
-import { CSSObject } from "@emotion/serialize/types/index";
+import { CSSObject } from "@emotion/react/types/index";
 import { RefObject } from "react";
 import NormalisedURLPath from "../../normalisedURLPath";
 import { API_RESPONSE_STATUS, SUCCESS_ACTION } from "./constants";